(fix) code style change
