+++ hummingbot/cli/config/config_validators.py
@@ -2,11 +2,11 @@ from os.path import (
     isfile,
     join,
 )
+from hummingbot.cli.utils.symbol_fetcher import SymbolFetcher
 from hummingbot.cli.settings import (
     EXCHANGES,
     STRATEGIES,
     CONF_FILE_PATH,
-    symbol_fetcher,
 )
 
 
@@ -26,6 +26,7 @@ def is_path(value: str) -> bool:
 def is_valid_market_symbol(market: str, value: str) -> bool:
     # Since symbol validation and autocomplete are UI optimizations that do not impact bot performances,
     # in case of network issues or slow wifi, this check returns true and does not prevent users from proceeding,
+    symbol_fetcher: SymbolFetcher = SymbolFetcher.get_instance()
     if symbol_fetcher.ready:
         market_symbols = symbol_fetcher.symbols.get(market, [])
         return value in symbol_fetcher.symbols.get(market) if len(market_symbols) > 0 else True
