+++ hummingbot/cli/utils/ethereum.py
@@ -1,0 +1,20 @@
+import logging
+from web3 import Web3
+
+
+def check_web3(ethereum_rpc_url: str) -> bool:
+    try:
+        w3: Web3 = Web3(Web3.HTTPProvider(ethereum_rpc_url, request_kwargs={"timeout": 2.0}))
+        ret = w3.isConnected()
+    except Exception:
+        ret = False
+
+    if not ret:
+        if ethereum_rpc_url.startswith("http://mainnet.infura.io"):
+            logging.getLogger().warning("You are connecting to an Infura using an insecure network protocol "
+                                        "(\"http\"), which may not be allowed by Infura. "
+                                        "Try using \"https://\" instead.")
+        if ethereum_rpc_url.startswith("mainnet.infura.io"):
+            logging.getLogger().warning("Please add \"https://\" to your Infura node url.")
+    return ret
+
